Sample AirBnB reviews for Discovery Service
A robust documentation and code set regarding the Discovery service will soon appear in this repository. For now, we encourage you to work with your instructor and for the joy of it, we have here over 100 AirBnB review regarding Manhattan apartments. Let's ingest these and do a a bit of discovery.